refactor: Move tested internals from helpers to the internal test file.
